64-bit build fixes
Former-commit-id: 7038e0e8b0b75204882f04e0fe6f872bbc68cd82

+++ Code/Dependencies/Prebuilt/Win64/include/G3D/XML.h
@@ -128,7 +128,7 @@ public:
 
     /** Attribute table size; zero for a TAG */
     int numAttributes() const {
-        return m_attribute.size();
+        return (int)m_attribute.size();
     }
 
     /** Return child \a i.  Children are nested tags and the unquoted
